fixup! refactor(devtools): improve components tab side pane UI
