+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/directive-explorer/property-tab/property-view/BUILD.bazel
@@ -11,6 +11,7 @@ _STYLE_SRCS = [
     "property-view-body.component.scss",
     "property-view-header.component.scss",
     "property-view-tree.component.scss",
+    "dependency-viewer.component.scss",
 ]
 
 _STYLE_LABELS = [
@@ -32,6 +33,7 @@ _STYLE_LABELS = [
 ng_module(
     name = "property-view",
     srcs = [
+        "dependency-viewer.component.ts",
         "property-editor.component.ts",
         "property-preview.component.ts",
         "property-tab-body.component.ts",
@@ -48,6 +50,7 @@ ng_module(
         "property-preview.component.html",
         "property-editor.component.html",
         "property-tab-body.component.html",
+        "dependency-viewer.component.html",
     ] + _STYLE_LABELS,
     deps = [
         "//devtools/projects/ng-devtools/src/lib/application-environment",
